+++ extensions/material-helper/src/extension.ts
@@ -13,6 +13,8 @@ import { registerDebugCommand } from './utils/debugMaterials';
 import { createComponentsTreeView } from './views/componentsView';
 import { createPagesTreeView } from './views/pagesView';
 import mtopAutoComplete from './mtopAutoComplete';
+import importAutoComplete from './importAutoComplete';
+import propTypesAutoComplete from './propTypesAutoComplete';
 
 const { window, ViewColumn } = vscode;
 
@@ -121,10 +123,14 @@ export function activate(context: vscode.ExtensionContext) {
   mtopAutoComplete();
   registerComponentDocSupport();
   autoFillContent();
-  
+
   // views
   createComponentsTreeView(context);
   createPagesTreeView(context);
+
+  importAutoComplete();
+  // help user complete React Component's propTypes; 
+  propTypesAutoComplete();
 }
 
 export function deactivate() { }