+++ extensions/material-helper/src/propTypesAutoComplete/getDefinitions.ts
@@ -1,0 +1,10 @@
+import * as vscode from 'vscode';
+
+export default async (uri: vscode.Uri, position: vscode.Position) => {
+  return await vscode.commands.executeCommand<vscode.Location[]>(
+    'vscode.executeDefinitionProvider',
+    uri,
+    position,
+  ) || [];
+};
+