+++ packages/doctor/CHANGELOG.md
@@ -1,5 +1,9 @@
 # changelog
 
+## 0.2.1
+
+- Support custom config set extends and plugins.
+
 ## 0.2.0
 
 - Remove codemod CLI usage.