+++ packages/project-service/CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## 0.1.2
+
+- fix: fail to execute getProjectDefInfo function in external
+- fix: check latest package version too slow
+
 ## 0.1.1
 
 - fix: rax-spa project targets null problem. [#861](https://github.com/appworks-lab/pack/issues/861)